+++ src/Jadob/Security/Supervisor/EventListener/SupervisorListener.php
@@ -5,10 +5,11 @@ declare(strict_types=1);
 namespace Jadob\Security\Supervisor\EventListener;
 
 use Jadob\Core\Event\BeforeControllerEvent;
+use Jadob\Core\RequestContext;
 use Jadob\Security\Auth\Exception\AuthenticationException;
 use Jadob\Security\Auth\Exception\InvalidCredentialsException;
 use Jadob\Security\Auth\Exception\UserNotFoundException;
-use Jadob\Security\Auth\UserStorage;
+use Jadob\Security\Auth\IdentityStorage;
 use Jadob\Security\Supervisor\RequestSupervisor\RequestSupervisorInterface;
 use Jadob\Security\Supervisor\Supervisor;
 use Psr\EventDispatcher\ListenerProviderInterface;
@@ -29,20 +30,18 @@ class SupervisorListener implements ListenerProviderInterface
     protected Supervisor $supervisor;
 
     /**
-     * @var UserStorage
+     * @var IdentityStorage
      */
-    protected UserStorage $userStorage;
+    protected IdentityStorage $identityStorage;
 
     /**
-     * SupervisorListener constructor.
-     *
      * @param Supervisor $supervisor
-     * @param UserStorage $userStorage
+     * @param IdentityStorage $identityStorage
      */
-    public function __construct(Supervisor $supervisor, UserStorage $userStorage)
+    public function __construct(Supervisor $supervisor, IdentityStorage $identityStorage)
     {
         $this->supervisor = $supervisor;
-        $this->userStorage = $userStorage;
+        $this->identityStorage = $identityStorage;
     }
 
     /**
@@ -69,11 +68,10 @@ class SupervisorListener implements ListenerProviderInterface
         /**
          * There is nothing to do when there is no supervisor available
          */
-        if($requestSupervisor === null) {
+        if ($requestSupervisor === null) {
             return $event;
         }
 
-        $this->userStorage->setCurrentProvider(get_class($requestSupervisor));
 
         //At first, handle stateless
         if ($requestSupervisor->isStateless()) {
@@ -85,7 +83,7 @@ class SupervisorListener implements ListenerProviderInterface
             return $event;
         }
 
-        $response = $this->handleNonStatelessRequest($event->getRequest(), $requestSupervisor);
+        $response = $this->handleNonStatelessRequest($event->getContext(), $requestSupervisor);
 
         if ($response !== null) {
             $event->setResponse($response);
@@ -125,12 +123,14 @@ class SupervisorListener implements ListenerProviderInterface
             return $supervisor->handleAuthenticationFailure($exception, $request);
         }
 
-        $this->userStorage->setUser($user, get_class($supervisor));
-        $supervisor->handleAuthenticationSuccess($request, $user);
+        $this->identityStorage->setUser($user, $request->getSession(), get_class($supervisor));
+        return $supervisor->handleAuthenticationSuccess($request, $user);
     }
 
-    protected function handleNonStatelessRequest(Request $request, RequestSupervisorInterface $supervisor): ?Response
+    protected function handleNonStatelessRequest(RequestContext $context, RequestSupervisorInterface $supervisor): ?Response
     {
+        $request = $context->getRequest();
+
         //1. Check if this is an authentication attempt:
         if ($supervisor->isAuthenticationRequest($request)) {
             try {
@@ -160,14 +160,12 @@ class SupervisorListener implements ListenerProviderInterface
                 return $supervisor->handleAuthenticationFailure($exception, $request);
             }
 
-            $this->userStorage->setUser($user, get_class($supervisor));
+            $this->identityStorage->setUser($user, $request->getSession(), get_class($supervisor));
             return $supervisor->handleAuthenticationSuccess($request, $user);
         }
 
-        /**
-         * Gets User from session storage.
-         */
-        $userFromStorage = $this->userStorage->getUser(get_class($supervisor));
+        $userFromStorage = $this->identityStorage->getUser($request->getSession(), get_class($supervisor));
+        $context->setUser($userFromStorage);
 
         /**
          * Case #1: User is logged in, nothing to do
@@ -194,6 +192,7 @@ class SupervisorListener implements ListenerProviderInterface
         }
 
 
+        return null;
         //3. User is not logged in, but supervisor allows unauthenticated user to enter
         //4. User is not logged in and supervisor wants user to be authenticated
         //5. User is logged in, there is nothing to do